+++ src/event.ts
@@ -1,4 +1,5 @@
 import {Log} from "./log";
+import 'core-js';
 
 export const EVENTS = {
     LOAD: 'load',